nav-bottom追加 #35[WIP], 認証周り順番変更
