+++ front/app/src/main/AndroidManifest.xml
@@ -31,10 +31,6 @@
             android:name=".WalkThroughActivity"
             android:label="@string/app_name"
             android:theme="@style/AppTheme.NoActionBar">
-            <intent-filter>
-                <action android:name="android.intent.action.MAIN" />
-                <category android:name="android.intent.category.LAUNCHER" />
-            </intent-filter>
         </activity>
         <activity
             android:name=".InitialSettingsActivity"
