+++ front/app/src/main/res/layout/activity_main.xml
@@ -23,4 +23,12 @@
 
     <include layout="@layout/content_main" />
 
+    <com.google.android.material.bottomnavigation.BottomNavigationView
+        android:layout_height="wrap_content"
+        android:layout_width="match_parent"
+        android:background="@color/backgroundMain"
+        android:id="@+id/bottom_nav"
+        app:menu="@menu/menu_bottom"
+        android:layout_gravity="bottom"/>
+
 </androidx.coordinatorlayout.widget.CoordinatorLayout>
