+++ front/app/src/main/res/menu/menu_bottom.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/main"
+        android:orderInCategory="100"
+        android:title="main" />
+    <item
+        android:id="@+id/ranking"
+        android:orderInCategory="200"
+        android:title="ranking"/>
+</menu>
